Eliminar importes no utilizados y metodos

+++ VitalFit_Service/src/main/java/com/tfg/vitalfit/service/repository/ObservacionesRepository.java
@@ -3,5 +3,5 @@ package com.tfg.vitalfit.service.repository;
 import com.tfg.vitalfit.service.entity.Observaciones;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ObservacionesRepository extends CrudRepository<Observaciones, Integer> {
+public interface ObservacionesRepository extends CrudRepository<Observaciones, Long> {
 }
