+++ VitalFit_Service/src/main/java/com/tfg/vitalfit/service/repository/OperacionesRepository.java
@@ -3,5 +3,5 @@ package com.tfg.vitalfit.service.repository;
 import com.tfg.vitalfit.service.entity.Operaciones;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OperacionesRepository extends CrudRepository<Operaciones, Integer> {
+public interface OperacionesRepository extends CrudRepository<Operaciones, Long> {
 }
