+++ VitalFit_Service/src/main/java/com/tfg/vitalfit/service/repository/PesosRepository.java
@@ -3,5 +3,5 @@ package com.tfg.vitalfit.service.repository;
 import com.tfg.vitalfit.service.entity.Pesos;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PesosRepository extends CrudRepository<Pesos, Integer> {
+public interface PesosRepository extends CrudRepository<Pesos, Long> {
 }
